fix: 🚑️ change dashboard column from json to long text
This deserves a better explanation. We have been facing memory issues with our database for months.  After investigations, we discovered that the JSON column, for some reason, causes MySQL to use gigabytes of memory to handle and makes the OOM Killer very upset (which is why itemdb crashes—especially when I am away from the computer).  Since we don't perform any operations on this column, there is no reason for it to be JSON, so we will just migrate it to text.

[skip ci]